+++ ipcache/ipcache_test.go
@@ -1,0 +1,23 @@
+package ipcache_test
+
+import (
+	"testing"
+	"time"
+
+	"github.com/m-lab/traceroute-caller/ipcache"
+)
+
+func TestRecentIPCache(t *testing.T) {
+	var tmp ipcache.RecentIPCache
+	ipcache.IP_CACHE_TIME_SECONDS = 20
+	tmp.New()
+	tmp.Add("1.2.3.4")
+	if !tmp.Has("1.2.3.4") {
+		t.Error("cache not working correctly")
+	}
+
+	time.Sleep(22 * time.Second)
+	if tmp.Has("1.2.3.4") {
+		t.Error("cache not expire correctly")
+	}
+}
